requirements.txt
torch==1.7.1
torchvision==0.8.2
matplotlib==3.1.3
seaborn==0.11.0
boilr==0.5.1
PyYAML==5.3.1
scikit-learn
h5py==2.8.0
opencv-python
tensorboard
wandb